Change size type to size_t

+++ FIFO.h
@@ -116,7 +116,7 @@ extern "C" {
 	 * structure that contains the buffer information (fifo_t). NULL is returned if
 	 * something fails.
 	 */
-	fifo_t fifo_create(uint16_t count, uint16_t size);
+	fifo_t fifo_create(uint16_t count, size_t size);
 
 	/**
 	 * @brief Creates a statically allocated FIFO buffer
